Code clean up for warehouse system
Changes made:
1) Deleted GetProductBalance() from WarehouseController.cs
2) Transfered UpdateProductBalance() to ProductBalanceController.cs
3) Deleted WarehouseController.cs (it had no more methods)
4) Merged GetOrdersByWareHouse() into GetOrders() to reduce code duplication
